+++ tests/shape_test.py
@@ -3,6 +3,7 @@
 # @file
 
 import numpy as np
+import pytest
 import scipp as sc
 from .common import assert_export
 
@@ -20,6 +21,59 @@ def test_fold():
     assert_export(sc.fold, x=x, dims=['x', 'y'], shape=[2, 3])
     assert_export(sc.fold, x=da, sizes={'x': 2, 'y': 3})
     assert_export(sc.fold, x=da, dims=['x', 'y'], shape=[2, 3])
+
+
+def test_fold_size_minus_1_variable():
+    x = sc.array(dims=['x'], values=np.arange(6.0))
+    assert sc.identical(sc.fold(x, dim='x', sizes={
+        'x': 2,
+        'y': 3
+    }), sc.fold(x, dim='x', sizes={
+        'x': 2,
+        'y': -1
+    }))
+    assert sc.identical(sc.fold(x, dim='x', sizes={
+        'x': 2,
+        'y': 3
+    }), sc.fold(x, dim='x', sizes={
+        'x': -1,
+        'y': 3
+    }))
+
+
+def test_fold_size_minus_1_data_array():
+    x = sc.array(dims=['x'], values=np.arange(6.0))
+    da = sc.DataArray(x)
+    assert sc.identical(sc.fold(da, dim='x', sizes={
+        'x': 2,
+        'y': 3
+    }), sc.fold(da, dim='x', sizes={
+        'x': 2,
+        'y': -1
+    }))
+    assert sc.identical(sc.fold(da, dim='x', sizes={
+        'x': 2,
+        'y': 3
+    }), sc.fold(da, dim='x', sizes={
+        'x': -1,
+        'y': 3
+    }))
+
+
+def test_fold_raises_two_minus_1():
+    x = sc.array(dims=['x'], values=np.arange(6.0))
+    da = sc.DataArray(x)
+    with pytest.raises(sc.DimensionError):
+        sc.fold(x, dim='x', sizes={'x': -1, 'y': -1})
+        sc.fold(da, dim='x', sizes={'x': -1, 'y': -1})
+
+
+def test_fold_raises_non_divisible():
+    x = sc.array(dims=['x'], values=np.arange(10.0))
+    da = sc.DataArray(x)
+    with pytest.raises(ValueError):
+        sc.fold(x, dim='x', sizes={'x': 3, 'y': -1})
+        sc.fold(da, dim='x', sizes={'x': -1, 'y': 3})
 
 
 def test_flatten():